requirements.txt
pandas
numpy
scikit-learn
openai
python-dotenv
pydantic
markdown-pdf